app/requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.3
Django==3.2.5
django-cors-headers==3.7.0
djangorestframework==3.12.4
googlemaps==4.4.7
idna==3.2
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1
urllib3==1.26.6
polyline==1.4.0
gunicorn==20.0.4
psycopg2-binary==2.8.5